gnuplot has been integrated so that you can use it to visualize data which we previously could only visualize with grace.  I've managed to statically link gnuplot in windows, test it, and generally it seems to work.
Currently the gnuplot executables are packaged within a single directory, but additional gnuplot documentation and demos are excluded from the package.  I intially attempted to include the entire directory structure, but this produced some extremely weird results.  Although the original gnuplot directory structure was only two levels deep, when it was called on my system with the java static executable linking routines, a very broken directory structure that was at least six levels deep was created.  This hindered the correct execution of the program.

There is a linux script included which should presumably run gnuplot.  I'm programming in windows right now, and I attempted to test this script on a remote linux system using a local copy of cygwin to allow X11 packet forwarding.  gnuplot started and seemed to display the data in question correctly.  I may need to have access to a local linux box to test this further, or maybe Bruce/Tim can check out whether this works on their computer.

There is currently no mac support.