Tim and I managed to get gnuplot working in a pretty decent fashion from linux.  Currently the bash script checks for gnome-terminal, and then xterm on the user's path.  As long as one of those console programs is present, gnuplot should start up and keep an interactive console open for the user to play with the data.  At this point, using commands such as
set style to lines
replot

Doesn't work really well.  The user has to take notice of the filename printed on the chart and do something along the lines of 

plot "<some obnoxiously long path to the .dat file>" with lines

in order to change the way the graph is represented.  So it goes.  Given the way gnuplot works in linux, I'm not sure I can do a lot better than this.