edgelist -> NWB converter has been cleaned up pretty well.
Right now it handles edgelists of two-column and three-column form in a pretty decent manner.

Third column is always assumed to be weight, and is treated as an int.

Issues:
multi-word labels in the edgelist file such as "hello gene 1"  are not treated as tokens.  All labels/IDs in the edgelist should be one word long.