The NAT now does the following:
lists all self-loops (though the information isn't very helpful)
lists all parallel edges (once again, the information isn't very helpful)
prints the # of weakly connected components
prints the size of the largest component

if there are self-loops or parallel edges and the graph claims to be undirected, a warning is printed to the screen.

if there are no self-loops or parallel edges, the density is calculated and displayed.