Now, if Aggregation Function File is not specified, the merging operation can work properly. Also include Micah's fix in isSameValue() method. It seems that with these fixes, the plugin now can handle merging paper-citation network with or without additional nodes attributes properly.
The problem left is when the network has nodes attributes, no matter the 
merging table has those attributes or not, if the user didn't specify a 
merging aggregation function file, the network structure did get updated 
(nodes and edges were merged), but the node attribute values remain without
aggregation. Therefore, those attributes values are not 100% accurate and 
can't be used.  
The proper handling should be that if not specify a merging aggregation 
function file, the updated network should automatically drop all node attributes 
and edge attributes.