Added midpoints to both kinds of LegendComponents.  Now in addition to showing visualization references for the minimum and maximum of the (scalable) input data, we show such a reference for the apparent midpoint (average of apparent minimum and apparent maximum), which is likewise labeled with the number that would be (or is) associated with that visual.  For the CircleAreaLegend, this is a third circle whose area is the mean of the extreme circles.  For ColorLegend, this is a third label at the midpoint of the color gradient.
In the case of ColorLegend, added small ticks on the gradient corresponding to the locations of the minimum, midpoint, and maximum.

LegendComponents now align by their key labels rather than their extrema labels.  This means that "Circle Size\nPopulation" and "Circle Color\nGDP", for example, are on the same horizontal rule (where before the numeric labels for the minimum and maximum had been the alignment anchors).

Added fine-print headers and footers to the page to show information like the author name, input data table's label, date & time, and attribution to CNS.  This meant altered the PostScript's BoundingBox DSC comment to be more lax, as these things would have been outside the original bounds, which had been calculated using only the map and legend.

Changed the page dimensions from that of US Letter portrait to that of US Letter landscape.  The mechanism for this uses a PostScript operator that is forbidden in Encapsulated PostScript: setpagedevice.  We must find another way to specify landscape-friendly page dimensions that will be recognized and respected by all the common PostScript renderers, or we must drop any claim to EPS-hood.

Added a page title.

Tweaked the polar crop region sizes to better agree with the scope of common global Mercator projections.

Added more metadata to the DSC prolog.

Changed the method used by LegendComponents to pretty print doubles for their numeric labels.  This is a fine issue as there are conflicting pressures between user-friendliness and accuracy.  The current method is not good enough and will need to be replaced.