For a SciPolicy mini-sprint story.
Until now, Geo Maps were laid out by fixing page dimensions and then scaling the map to fit inside.

Now we lay out the page around the map.  In particular, the page width and map width are constant (see utility.Constants).  The map height is the height necessary to fix the aspect ratio of the displayed map to that of the shapefile feature data.  The page height is the height necessary to contain the page header, the map, the legend, and the page footer.