For Sci2 mini-sprint.
Updated how double values are pretty-printed by the LegendComponents.  Previously we could get values like "48,687   254,050.06   1,325,640".  Now we get "48,687   254,050   1,325,640".  The approximation is done using a DecimalFormat with only as many digits right of the radix as necessary to represent the values precisely (as measured by relative difference) and distinctly (so that "0.00001   0.00002   1" does not come out as "0   0   1").  Digits left of the radix are never modified.  When the DecimalFormat would yield "-0", instead give "0".  This functionality is in legend/UnsignedZeroDecimalFormat.java.

Reviewed by Chintan.