In the region-coloring style, we read region names (like "Indiana" or "United States") from the input table and try to find them in the shapefile for coloring.  Previously this was case-sensitive.  Now we will match "United States" (as it is in the shapefile) with "UNITED STATES" (from the user's table).
Commented out "setpagedevice" for the time being.  We're already generating roughly A4-ish or U.S. Letter-ish page sizes.  Removing setpagedevice, which is a PostScript command not permitted in Encapsulated PostScript, may aid consistent rendering among, for present purposes, GhostView, Adobe Distiller, and Mac Preview.

Added an apparently non-standard EPS metadatum to the DSC prolog.  GhostScript recognizes "%%Orientation: Landscape", and since we are aiming for GhostScript compatibility at the moment, this is our preferred method.  Consequently, note that we have swapped the coordinates in the BoundingBox setting (we used to set a box in a landscape orientation, now we set a box in portrait orientation and ask for the file to be landscape oriented).  The final third of this orientation redesign is the addition of a global rotate by 90 degrees followed immediately by a global translate to slide the image onto the page.

Halved the line width of feature/region borders from 0.8 to 0.4.

Bumped up the south crop height from 11 degrees to 12 (these are latitudinal away from the pole at 0 degrees latitude).

We have previously handled the drawing of feature geometry coordinates across map projection interruptions at the meridian opposite the central meridian.  This did not handle the case when users (against recommendation in both the algorithm parameters description for the projection and the documentation) used conic projections at the global scale (that is, with the Countries shapefile option).  Conic projections, unlike the other that we have, will produce ugly lines in the presence of the polar interruptions.  I've added a hacky check to see whether an unusually long straight line would be drawn for a feature, and in that case, we do not draw it.  See FeaturePrinter.