More SOM.
Dropped in MPI setup and Allreduce on numerators and denominators.

Cleaned up comments a bit.

Added another function to load training vectors.  This one expects a sparse file format (each row gives indices to ones in the binary-valued vector that it represents).  Renamed the older function to reflect dense usage.