Changed how the number of timesteps is derived from the user-specified number of training steps, number of jobs, and number of timesteps between updates/synchs (that is, epoch length).
Previously, depending on given said parameters, the code might perform a series of "full" epochs (that is, those whose actual length was as desired) and finish with a "fractional" epoch (that is, with less than desired length).  For certain values of said parameters, this final epoch could practically be very short, meaning that the final codebook would poorly represent the training data (it being only a handful of linear interpolations over the previous fully-updated codebook).  While, granted, this should not be an issue when the Gaussian width interpolation has resulted in a small final value, it was conceptually undesirable.

Now the number of timesteps is derived (by rounding up) such that one should never have a fractional final epoch.  This may result in a total run length slightly longer than requested, but is conceptually sound.