Fixed two problems with all-zero vectors in the initial, random codebooks.  It had been possible for the Python script to generate an all-zero codebook vector, and all-zero vectors were not checked for by the codebook reader in the C++ code.  This was fatal for the cosine similarity metric, which divides by the norm.  Now the script never gives all-zero vectors and the C++ code checks for all-zero vectors at the initial load.  Trying to load one is considering an immediately fatal error.
Added MPI error handling.  After MPI::Init, we now set the error handler to throw exceptions.  We try/catch for these at all three significant MPI call sites and treat them as fatal errors.

Changed the number of training vectors used by each process for quantization error checking from the constant 100 to the number of nodes in the map (which in tests so far has been around 100 to 100000).  This is arbitrary and probably ought to be tweaked.