Changed extract-central-mesh-terms to actively shuffle the dictionary keys.  They were probably already randomized, but now we don't have to worry about any implementation oddities.
Added a Python script that takes a codebook and a sparsely-represented dataset and produces a visualization of the codebook with each node colored in grayscale according to how many of the given data vectors found it to be the best match (with cosine dissimilarity).

Added a Python script to color codebook nodes according to one dimension of their vectors.