Added calibrate.cpp, a spiritual equivalent to SOM_PAK's vcal which allows for sparsely-represented training data and for cosine dissimilarity.  For a given codebook, it accepts training data and outputs the best-matching units and minimal cosine dissimilarities.  On Feynman, splitting to 4 processes through MPI, it seems a 252x300 codebook with 2300 dimensions and 2.1 million training vectors will take about 13 hours to finish.
Added codebook_planes.py, which generates one PostScript file with one page for each of the first n planes of a codebook.

Added visual.py, an interactive script that accepts one training vector at a time and gives the row and column of the BMU in the codebook (as well as the vector's cosine dissimilarity from it).

Added codebook_hits.py, which is meant to find each BMU for a selection of training vectors and create a greyscale PostScript visualizing the proportion of BMU hitcounts.  Don't use this; it is much too slow (code kept in case something can be cannibalized later).

Added qerror.cpp, intended to be a clone of SOM_PAK's quantization error program but allowing for sparsely-representing training data and using cosine dissimilarity.  Unfinished and untested.