Made MergeClustersBySimilarity a little smarter and a lot faster by storing dissimilarities in the heap entries rather than calculating them on-the-fly in the heap's comparator.  Thanks Russell.
Added MergeClustersBySizeThenSimilarity, a variant of MergeClustersBySize.  Both scripts choose pairs of neighboring clusters to merge.  The older one first identifies a cluster with the globally minimal number of documents, then merges it with its nearest neighbor.  That initial identification is just a heap pop, so any ties (which appear to be frequent on the low-end in the data considered so far) are broken however the heap chooses.  The newer one, a bit smarter and a lot more deterministically, essentially breaks ties by minimizing the distance of the candidate clusters from their nearest neighbor.  Of the three MergeClusters programs presented so far, this one should be preferred for reproducibly merging codebook neurons into optimally discriminative clusters of a fixed minimum size.

Added cluster-vis.py (and supporting file clusterTemplate.eps).  This expects a file mapping neurons (identifying by their row,column coordinate in the codebook) to cluster IDs.  The script generates a PostScript file which draws one small, light gray circle for each neuron in the codebook and draws a little black line between neighboring neurons if and only if they are mapped to distinct cluster IDs.

Added document-count-distribution.py, a simple script accepting output from one of the MergeClusters programs which prints comma-separated values representing the frequency of cluster document populations.

Added pmid-to-neuron.py, which just simplifies and reformats output from calibrate.cpp into a file mapping neuron IDs (in the case of the present data, PubMed IDs) to the row,column coordinate of the best-matching neuron in the codebook.

Added neuron-to-cluster.py, which just simplifies and reformats output from one of the MergeClusters programs into a file mapping row,column neuron coordinates to the cluster ID specified in said output.