Added support for arbitrary float values in sparse training vectors.  This entailed adding a new training data file loader (which entailed tweaking some of the existing terminology for the older sparse-representation training vectors to be explicitly those which are "binary-valued") and removing an optimizing assumption from the cosine similarity method (where it had sufficed to calculate the norm of a sparsely-represented binary-valued training vector by sqrt'ing the size of the corresponding map, we must now actually calculate the norm).
We had previously optimized the code as intended for use with the Euclidean distance metric to calculate the norms of codebook vectors only once per epoch.  In fact, we recorded the square of these norms as that was what our formula ultimately used.  With cosine similarity, however, we ultimately do want the norm proper, and so the code has now been tweaked to record proper(/unsquared) norms.  This removes the previous need for a call to sqrt at each invocation of the cosine similarity method.

Inlined a function with an outdated and misleading name.

Commented out the Euclidean metric stuff.

Tweaked some variable/method names and added some comments on further optimization.

This represents the state of the SOM code at the beginning of the NIH maps job.