Adding runner algorithm for the "exact" epidemic simulator, edu.iu.simulator.core.exact.
Restructured existing runner for the "single-population" epidemic simulator (edu.iu.simulator.core.singlepopulation) to take advantage of nearly identical code across runners.  This code is in EpidemicSimulatorAlgorithm and its Factory.  Renamed a lot of things to remove specific references to SPEMShell.

Renamed simulator.core plugins in accordingly.

For now, the single-population runner is just an extension of the exact runner except it:
* Refers to the single-population simulator core algorithm PID
* Adds an algorithm parameter for the random number generator seed
* Adds an algorithm parameter for the number of runs (this is unnecessary in the exact case; each run would produce the same output)

Added CompartmentNameOrdering to ensure that dynamic algorithm parameters corresponding to compartments are listed in a fixed order across different invocations of the algorithm or of the tool.  An effort is made to give orders like S, E, Ia, It, R.

Edited menu and EpiC build files to add/update exact and single-population runners.

Reviewed by Chin Hua.