JIRA NETE-2  This splits the temporalbargraph and gives it a version for the web
The web version is a cleaned version of the previous temporalbargraph.  It does not allow the user to specify a page orientation.  It also have a much 'cleaner' look for displaying for webpages.  The code now uses stringtemplating to display the postscript, which is much more readable.  There are still improvements needed such as making the integration with the former model tighter, adding string templates for the regular temporal bar graph, and using joda time.

Reviewed by Joseph.