The code at a high level only adds the feature of users can specify a category to color by.  NETE-20 Add color coding to Temporal bar graph.  There is still no legend mockup for the color coding so one is not present in this code.
At a lower level, the code base has been throughly cleaned.  The two algorithms now inherit from the same abstract classes and both use string templating (which they also share) to add elements to a page.  It would be much easier for this code to be changed to render something other than postscript.

Reviewed by Joseph.