!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!! This is broken because the .classpath is missing.  I will commit in next so I can pass the checksum.
r5637 broke the nightly build.  I am trying to fix that issue.

The problem revolves around java 1.6.  In this plugin, the build is not using 1.6, but in others it will use it.

I will add in another .classpath file in the next commit so I can pass the checksum match for