Repositioning the single color legend in the region-coloring mode.  Now it's always along the far left margin.
As of this commit I'm noticing two major bugs that have been introduced somewhere along the way.  First, region coloring seems to be completely broken.  Second, the Mercator and Winkel Tripel transforms are broken (at least for the world map).  The Mercator projection is distorted in some peculiar way and Winkel Tripel is throwing an exception relating to its central meridian parameter.