Now italicizing map name in the how-to-read text.  The method for this is brittle and hard to understand.  To avoid passing too many primitives around, the HowToRead object is now constructed at the Algorithm level and passed down as an Optional<HowToRead>.
By request the how-to-read text now also includes the number of regions present in the base map.  This is calculated as the number of unique feature names present in the shapefile.

The default circle area, used when circle area coding is disabled, is by request now much, much smaller.

The title "Legend" normally printed on the "legendarium" is now suppressed when it contains no legends.  This can happen when no available codings are enabled.

New, naive "showLines" PostScript utility.  Similar to existing "showAndNewLine" but uses moveto instead of rmoveto.