All entries in both the discipline and subdiscipline distribution analyses are now counting journal occurrences rather than unique journals (added forwarding method on JournalDataset to Multiset.count for this purpose).  This method of counting agrees with past results.  Previously the "Unmapped" row of the subdiscipline analysis and all rows of the discipline analysis were not being counted with multiplicity like this.
Introduced AbstractTabularAnalysis ("An analysis of a sequence of bins in terms of their values over a set of typed columns.") for common functionality in the discipline and subdiscipline analyses.  Reworked SubdisciplineAnalysis in terms of this and created DisciplineAnalysis similarly.

Introduced Column ("A typed attribute, as of an AbstractTabularAnalysis or a Row.") and Row ("A type-safe heterogeneous container for mapping each of a set of typed Columns to a value of the appropriate type.").  These can be used to prepare tabular data for representation as a prefuse Table in a comparatively type-safe way.

Whether to make use of the canonical journal form lookup and whether to include the discipline and subdiscipline distribution analyses in the journal algorithm's output data are now controlled by global booleans on JournalsMapAlgorithm.  This is in anticipation of merging this branch back into trunk with these values set to false.