Fixed error message date and broken URL. Resolves Jira SCISQUARED-1205
Error message example date was being taken from a sample of the user's
input data, and going through some data type as well as formatting
changes. Was able to simplify by using a current date as the example
date and formatting it with the user-definied format immediately.

The java docs link was broken. Updated it to an active link.